feat: continue work on fix type of check for ranges rule
FossilOrigin-Name: 6218e445b78f3ebe1e4728299605c8e1adf527eb738ae966f539de5280b47461

+++ src/rules/ranges.nim
@@ -74,7 +74,9 @@ ruleConfig(ruleName = "ranges",
 
 type FileLine = string
 
-var fileContent: seq[string] = @[]
+var
+  fileContent: seq[string] = @[]
+  needSave: bool = false
 
 checkRule:
   initCheck:
@@ -92,9 +94,13 @@ checkRule:
   checking:
     try:
       if node.kind == nkIdent and $node == "..":
-        let rangeLine: FileLine = fileContent[node.info.line.Natural - 1]
-        setResult(checkResult = rangeLine[rangeLine.find(sub = "..") - 1] ==
-            ' ', positiveMessage = positiveMessage,
+        let
+          rangeLine: FileLine = fileContent[node.info.line.Natural - 1]
+          checkResult: bool = rangeLine[rangeLine.find(sub = "..") - 1] == ' '
+        if rule.ruleType == fix and not needSave and (not checkResult or (
+            checkResult and rule.negation)):
+          needSave = true
+        setResult(checkResult = checkResult, positiveMessage = positiveMessage,
             negativeMessage = negativeMessage, ruleData = $node.info.line,
             node = node, params = [$node.info.line, hasMessage &
             " spaces between start and end of the range"])
@@ -102,24 +108,25 @@ checkRule:
       rule.amount = errorMessage(text = messagePrefix & "can't check file '" &
           rule.fileName & ". Reason: ", e = getCurrentException())
   endCheck:
-    discard
+    if rule.ruleType == fix and needSave:
+      let newFileName: FilePath = rule.fileName & ".bak"
+      try:
+        moveFile(source = rule.fileName, dest = newFileName)
+        let newFile: File = open(filename = rule.fileName, mode = fmWrite)
+        for index, line in fileContent:
+          newFile.writeLine(x = line)
+        newFile.close
+      except OSError, IOError, Exception:
+        discard errorMessage(text = "Can't fix file '" &
+            rule.fileName & ". Reason: ", e = getCurrentException())
+        try:
+          removeFile(file = rule.fileName)
+        except OSError:
+          discard
+        try:
+          moveFile(source = newFileName, dest = rule.fileName)
+        except IOError, OSError, Exception:
+          discard
 
 fixRule:
-
-  let newFileName: FilePath = rule.fileName & ".bak"
-  try:
-    moveFile(source = rule.fileName, dest = newFileName)
-    let newFile: File = open(filename = rule.fileName, mode = fmWrite)
-    newFile.close
-  except OSError, IOError, Exception:
-    discard errorMessage(text = "Can't fix file '" &
-        rule.fileName & ". Reason: ", e = getCurrentException())
-    try:
-      removeFile(file = rule.fileName)
-    except OSError:
-      discard
-    try:
-      moveFile(source = newFileName, dest = rule.fileName)
-    except IOError, OSError, Exception:
-      discard
-    return false
+  discard
